Split layers into different headers and compile units
