+++ canopen_motor_node/CMakeLists.txt
@@ -22,7 +22,22 @@ find_package(Boost REQUIRED
 find_package(PkgConfig)
 pkg_check_modules (MUPARSER REQUIRED muparser)
 
-catkin_package()
+catkin_package(
+  INCLUDE_DIRS
+    include
+  LIBRARIES
+    canopen_motor
+  CATKIN_DEPENDS
+    canopen_402
+    canopen_chain_node
+    controller_manager
+    hardware_interface
+    joint_limits_interface
+    urdf
+  DEPENDS
+    Boost
+    MUPARSER
+)
 
 include_directories(
   include
@@ -30,14 +45,28 @@ include_directories(
   ${MUPARSER_INCLUDE_DIRS}
 )
 
-# canopen_motor_node
-add_executable(canopen_motor_node
-  src/control_node.cpp
+# canopen_motor
+add_library(canopen_motor
+  src/controller_manager_layer.cpp
+  src/handle_layer.cpp
+  src/motor_chain.cpp
   src/robot_layer.cpp
 )
-target_link_libraries(canopen_motor_node
+target_link_libraries(canopen_motor
   ${catkin_LIBRARIES}
   ${MUPARSER_LIBRARIES}
+)
+add_dependencies(canopen_motor
+  ${catkin_EXPORTED_TARGETS}
+)
+
+# canopen_motor_node
+add_executable(canopen_motor_node
+  src/canopen_motor_chain_node.cpp
+)
+target_link_libraries(canopen_motor_node
+  canopen_motor
+  ${catkin_LIBRARIES}
 )
 add_dependencies(canopen_motor_node
   ${catkin_EXPORTED_TARGETS}
@@ -45,18 +74,24 @@ add_dependencies(canopen_motor_node
 
 install(
   TARGETS
+    canopen_motor
     canopen_motor_node
   ARCHIVE DESTINATION ${CATKIN_PACKAGE_LIB_DESTINATION}
   LIBRARY DESTINATION ${CATKIN_PACKAGE_LIB_DESTINATION}
   RUNTIME DESTINATION ${CATKIN_PACKAGE_BIN_DESTINATION}
 )
 
+install(DIRECTORY include/${PROJECT_NAME}/
+  DESTINATION ${CATKIN_PACKAGE_INCLUDE_DESTINATION}
+  FILES_MATCHING PATTERN "*.h"
+)
+
 if(CATKIN_ENABLE_TESTING)
   catkin_add_gtest(${PROJECT_NAME}-test_muparser
     test/test_muparser.cpp
-    src/robot_layer.cpp
   )
   target_link_libraries(${PROJECT_NAME}-test_muparser
+    canopen_motor
     ${catkin_LIBRARIES}
     ${MUPARSER_LIBRARIES}
   )
